+++ packages/common/src/search/_internal/hubSearchItems.ts
@@ -4,6 +4,8 @@ import {
   IHubSearchResult,
   IQuery,
 } from "../types";
+import { getOgcCollectionUrl } from "./hubSearchItemsHelpers/getOgcCollectionUrl";
+import { ISearchOgcItemsOptions } from "./hubSearchItemsHelpers/interfaces";
 import { searchOgcAggregations } from "./hubSearchItemsHelpers/searchOgcAggregations";
 import { searchOgcItems } from "./hubSearchItemsHelpers/searchOgcItems";
 
@@ -18,7 +20,9 @@ export async function hubSearchItems(
   query: IQuery,
   options: IHubSearchOptions
 ): Promise<IHubSearchResponse<IHubSearchResult>> {
-  return options.aggFields?.length
-    ? searchOgcAggregations(query, options)
-    : searchOgcItems(query, options);
+  if (options.aggFields?.length) {
+    return searchOgcAggregations(query, options);
+  }
+  const url = `${getOgcCollectionUrl(query, options)}/items`;
+  return searchOgcItems(url, query, options as ISearchOgcItemsOptions);
 }