+++ packages/common/src/search/_internal/hubSearchItemsHelpers/formatOgcItemsResponse.ts
@@ -1,20 +1,21 @@
 import { IQuery } from "../../types/IHubCatalog";
-import { IHubSearchOptions } from "../../types/IHubSearchOptions";
 import { IHubSearchResponse } from "../../types/IHubSearchResponse";
 import { getNextOgcCallback } from "./getNextOgcCallback";
-import { IOgcItemsResponse } from "./interfaces";
+import { IOgcItemsResponse, ISearchOgcItemsOptions } from "./interfaces";
 import { ogcItemToSearchResult } from "./ogcItemToSearchResult";
 import { ogcItemToDiscussionPostResult } from "./ogcItemToDiscussionPostResult";
 import { IHubSearchResult } from "../../types";
 
 export async function formatOgcItemsResponse(
   response: IOgcItemsResponse,
+  url: string,
   originalQuery: IQuery,
-  originalOptions: IHubSearchOptions
+  originalOptions: ISearchOgcItemsOptions
 ): Promise<IHubSearchResponse<IHubSearchResult>> {
   if (originalQuery.targetEntity === "discussionPost") {
     return formatDiscussionPostTargetEntityResponse(
       response,
+      url,
       originalQuery,
       originalOptions
     );
@@ -22,6 +23,7 @@ export async function formatOgcItemsResponse(
 
   return formatItemTargetEntityResponse(
     response,
+    url,
     originalQuery,
     originalOptions
   );
@@ -29,13 +31,19 @@ export async function formatOgcItemsResponse(
 
 async function formatDiscussionPostTargetEntityResponse(
   response: IOgcItemsResponse,
+  url: string,
   originalQuery: IQuery,
-  originalOptions: IHubSearchOptions
+  originalOptions: ISearchOgcItemsOptions
 ): Promise<IHubSearchResponse<IHubSearchResult>> {
   const formattedResults = await Promise.all(
     response.features.map((f) => ogcItemToDiscussionPostResult(f))
   );
-  const next = getNextOgcCallback(response, originalQuery, originalOptions);
+  const next = getNextOgcCallback(
+    response,
+    url,
+    originalQuery,
+    originalOptions
+  );
   const nextLink = response.links.find((l) => l.rel === "next");
 
   return {
@@ -48,8 +56,9 @@ async function formatDiscussionPostTargetEntityResponse(
 
 async function formatItemTargetEntityResponse(
   response: IOgcItemsResponse,
+  url: string,
   originalQuery: IQuery,
-  originalOptions: IHubSearchOptions
+  originalOptions: ISearchOgcItemsOptions
 ): Promise<IHubSearchResponse<IHubSearchResult>> {
   const formattedResults = await Promise.all(
     response.features.map((f) =>
@@ -60,7 +69,12 @@ async function formatItemTargetEntityResponse(
       )
     )
   );
-  const next = getNextOgcCallback(response, originalQuery, originalOptions);
+  const next = getNextOgcCallback(
+    response,
+    url,
+    originalQuery,
+    originalOptions
+  );
   const nextLink = response.links.find((l) => l.rel === "next");
 
   return {