+++ packages/common/src/search/_internal/hubSearchItemsHelpers/getSortByQueryParam.ts
@@ -1,14 +1,14 @@
-import { IHubSearchOptions } from "../../types/IHubSearchOptions";
+import { ISearchOgcItemsOptions } from "./interfaces";
 
 /**
  * @private
  * Serializes the sort options provided from the request options
  * object into a string that the OGC API can understand
  *
- * @param options IHubSearchOptions that contain sorting information
+ * @param options ISearchOgcItemsOptions that contain sorting information
  * @returns a serialized sort string
  */
-export function getSortByQueryParam(options: IHubSearchOptions) {
+export function getSortByQueryParam(options: ISearchOgcItemsOptions): string {
   const { sortField, sortOrder } = options;
   let result;
   if (sortField) {