+++ packages/common/test/search/_internal/hubSearchItems.test.ts
@@ -22,7 +22,10 @@ import {
 import { getQueryString } from "../../../src/search/_internal/hubSearchItemsHelpers/getQueryString";
 import { getOgcAggregationQueryParams } from "../../../src/search/_internal/hubSearchItemsHelpers/getOgcAggregationQueryParams";
 import { getQQueryParam } from "../../../src/search/_internal/hubSearchItemsHelpers/getQQueryParam";
-import { IOgcItem } from "../../../src/search/_internal/hubSearchItemsHelpers/interfaces";
+import {
+  IOgcItem,
+  ISearchOgcItemsOptions,
+} from "../../../src/search/_internal/hubSearchItemsHelpers/interfaces";
 import * as ogcItemToSearchResultModule from "../../../src/search/_internal/hubSearchItemsHelpers/ogcItemToSearchResult";
 import * as ogcItemToDiscussionPostModule from "../../../src/search/_internal/hubSearchItemsHelpers/ogcItemToDiscussionPostResult";
 import { formatOgcItemsResponse } from "../../../src/search/_internal/hubSearchItemsHelpers/formatOgcItemsResponse";
@@ -385,7 +388,7 @@ describe("hubSearchItems Module |", () => {
       };
 
       it("handles query", () => {
-        const options: IHubSearchOptions = {};
+        const options: ISearchOgcItemsOptions = {};
         const result = getOgcItemQueryParams(query, options);
         const expected = {
           filter: "((type=typeA))",
@@ -402,7 +405,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query and auth", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -426,7 +429,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query, auth and limit", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -451,7 +454,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query, auth, limit and startindex", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -477,7 +480,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query, auth, limit, startindex and q", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -506,7 +509,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query, auth, limit, startindex, q and sortBy", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -540,7 +543,7 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("handles query, auth, limit, startindex, q, flatten, fields and sortBy", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             authentication: {
               token: "abc",
@@ -698,14 +701,14 @@ describe("hubSearchItems Module |", () => {
 
     describe("getSortByQueryParam |", () => {
       it("returns undefined if no sortField is provided", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           sortOrder: "asc",
         };
         const result = getSortByQueryParam(options);
         expect(result).toBeUndefined();
       });
       it("handles sorting in descending order", () => {
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           sortField: "title",
           sortOrder: "desc",
         };
@@ -965,11 +968,12 @@ describe("hubSearchItems Module |", () => {
 
     describe("getNextOgcCallback", () => {
       const { getNextOgcCallback } = getNextOgcCallbackModule;
+      const url = "https://my-hub.com/api/search/v1/collections/all/items";
       const query: IQuery = {
         targetEntity: "item",
         filters: [],
       };
-      const options: IHubSearchOptions = { num: 1 };
+      const options: ISearchOgcItemsOptions = { num: 1 };
       const nextResponse: IHubSearchResponse<IHubSearchResult> = {
         total: 0,
         results: [],
@@ -986,7 +990,12 @@ describe("hubSearchItems Module |", () => {
       });
 
       it("returns an empty callback when no next link is present", async () => {
-        const callback = getNextOgcCallback(ogcItemsResponse, query, options);
+        const callback = getNextOgcCallback(
+          ogcItemsResponse,
+          url,
+          query,
+          options
+        );
         const callbackResult = await callback();
         expect(callbackResult).toBeNull();
         // NOTE: using `toHaveBeenCalled` throws a fatal error ONLY in Karma
@@ -998,6 +1007,7 @@ describe("hubSearchItems Module |", () => {
       it("returns a callback with modified options when next link is present", async () => {
         const callback = getNextOgcCallback(
           ogcItemsResponseWithNext,
+          url,
           query,
           options
         );
@@ -1009,7 +1019,7 @@ describe("hubSearchItems Module |", () => {
         const numCalls = searchOgcItemsSpy.calls.count();
         expect(numCalls).toBe(1);
         const callInfo = searchOgcItemsSpy.calls.first();
-        expect(callInfo.args).toEqual([query, { ...options, start: 2 }]);
+        expect(callInfo.args).toEqual([url, query, { ...options, start: 2 }]);
       });
     });
 
@@ -1017,11 +1027,12 @@ describe("hubSearchItems Module |", () => {
     // Ends up throwing errors ONLY in Karma. Figure out why that is and
     // test that `formatOgcItemsResponse` is delegating accordingly
     describe("formatOgcItemsResponse |", () => {
+      const url = "https://my-hub.com/api/search/v1/collections/all/items";
       const query: IQuery = {
         targetEntity: "item",
         filters: [],
       };
-      const requestOptions: IHubSearchOptions = {
+      const requestOptions: ISearchOgcItemsOptions = {
         include: [],
         requestOptions: {},
       };
@@ -1029,6 +1040,7 @@ describe("hubSearchItems Module |", () => {
       it("correctly handles when no next link is present", async () => {
         const formattedResponse = await formatOgcItemsResponse(
           ogcItemsResponse,
+          url,
           query,
           requestOptions
         );
@@ -1040,6 +1052,7 @@ describe("hubSearchItems Module |", () => {
       it("correctly handles when the next link is present", async () => {
         const formattedResponse = await formatOgcItemsResponse(
           ogcItemsResponseWithNext,
+          url,
           query,
           requestOptions
         );
@@ -1052,6 +1065,7 @@ describe("hubSearchItems Module |", () => {
       it("correctly handles discussion posts", async () => {
         const formattedResponse = await formatOgcItemsResponse(
           ogcDiscussionPostResponseWithNext,
+          url,
           { targetEntity: "discussionPost", filters: [] },
           requestOptions
         );
@@ -1124,7 +1138,7 @@ describe("hubSearchItems Module |", () => {
           return fakeResponse;
         };
 
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           site: `https://${siteHostname}`,
           requestOptions: {
             fetch: _fetch,
@@ -1146,8 +1160,29 @@ describe("hubSearchItems Module |", () => {
           return fakeResponse;
         };
 
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           site: hubApiUrl,
+          requestOptions: {
+            fetch: _fetch,
+            hubApiUrl,
+          },
+        };
+        await ogcApiRequest(hubApiUrl, queryParams, options);
+      });
+
+      it("does not append the ?target query param when there is no target site domain", async () => {
+        const fakeResponse = {
+          ok: true,
+          statusText: "200: Ok",
+          status: 200,
+          json: () => Promise.resolve(),
+        };
+        const _fetch: any = async (finalUrl: string) => {
+          expect(finalUrl).toBe(`${hubApiUrl}?type=CSV`);
+          return fakeResponse;
+        };
+
+        const options: ISearchOgcItemsOptions = {
           requestOptions: {
             fetch: _fetch,
             hubApiUrl,
@@ -1163,7 +1198,7 @@ describe("hubSearchItems Module |", () => {
           status: 404,
         };
         const _fetch: any = async () => fakeResponse;
-        const options: IHubSearchOptions = {
+        const options: ISearchOgcItemsOptions = {
           site: `https://${siteHostname}`,
           requestOptions: {
             fetch: _fetch,
@@ -1174,7 +1209,7 @@ describe("hubSearchItems Module |", () => {
           await ogcApiRequest(hubApiUrl, queryParams, options);
           expect(true).toBe(false);
         } catch (err) {
-          expect(err.message).toEqual("404: Not Found");
+          expect((err as any).message).toEqual("404: Not Found");
         }
       });
     });